split classes on packages
