+++ src/main/java/executor/service/model/configs/ThreadPoolConfig.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.configs;
 
 import java.util.Objects;
