--- src/main/java/executor/service/service/ProxySourcesClient.java
@@ -1,9 +1,0 @@
-package executor.service.service;
-
-import executor.service.model.ProxyConfigHolder;
-
-public interface ProxySourcesClient {
-
-    ProxyConfigHolder getProxy();
-
-}
