--- src/main/java/executor/service/service/impl/ProxySourcesClientImpl.java
@@ -1,11 +1,0 @@
-package executor.service.service.impl;
-
-import executor.service.model.ProxyConfigHolder;
-import executor.service.service.ProxySourcesClient;
-
-public class ProxySourcesClientImpl implements ProxySourcesClient {
-    @Override
-    public ProxyConfigHolder getProxy() {
-        return null;
-    }
-}
