+++ src/main/java/executor/service/service/proxy/ProxySourcesClient.java
@@ -1,0 +1,9 @@
+package executor.service.service.proxy;
+
+import executor.service.model.proxy.ProxyConfigHolder;
+
+public interface ProxySourcesClient {
+
+    ProxyConfigHolder getProxy();
+
+}
