+++ src/main/java/executor/service/service/proxy/selenium/ProxySourcesClientImpl.java
@@ -1,0 +1,11 @@
+package executor.service.service.proxy.selenium;
+
+import executor.service.model.proxy.ProxyConfigHolder;
+import executor.service.service.proxy.ProxySourcesClient;
+
+public class ProxySourcesClientImpl implements ProxySourcesClient {
+    @Override
+    public ProxyConfigHolder getProxy() {
+        return null;
+    }
+}
